move pyne source parameters
